% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtl_fun.R
\name{dtl_get_alpha_s}
\alias{dtl_get_alpha_s}
\title{get the alpha_s given significance level alpha}
\usage{
dtl_get_alpha_s(delta, n, t, rho, q, alpha)
}
\description{
get the alpha_s given significance level alpha
}
