% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtlcor_shiny.R
\name{dtl_shiny}
\alias{dtl_shiny}
\title{Shiny app of drop-the-losers (DTL) design}
\usage{
dtl_shiny(appname = "shiny")
}
\arguments{
\item{appname}{Default is "shiny". Do not change it since there is only
one shiny app in the package}
}
\value{
The shiny app of DTL design which includes three panels:(1) family-wise
type I error rate (FWER) given fixed rho; (2) correlation coefficient
boundary for rho (rho_s); (3) adjusted significance level in real
application (alpha_t). The first two panels show the graphs of change
of the FWER and the rho_s as the change of some related parameters. The
corresponding tables of the graphs are also shown. In the last
panel, the table of the significance levels alpha_s based on all possible
values of response rate q and hazard ratio of responders and
non-responders gamma and the resulting minimum or called adjusted
significance level are shown.
}
\description{
Interactive shiny app of drop-the-losers (DTL) design
}
\examples{
# run dtl_shiny()


}
