% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtl_fun.R
\name{dtl_tier_the}
\alias{dtl_tier_the}
\title{get the theoretical FWER given fixed correlation coefficient}
\usage{
dtl_tier_the(delta, n, t, rho, q, alpha_s)
}
\description{
get the theoretical FWER given fixed correlation coefficient
}
